feat: enable webapps BackupController even when not standalone and rename backups
Backup files are renamed using camunda_webapps_ instead of
camunda_operate
closes #25481
